Replace custom chai assertions with custom helpers
