--- e2e/support/properties/index.ts
@@ -1,1 +1,0 @@
-import "./scrollbars";
